+++ packages/atomic/storybookUtils/common/render-component.tsx
@@ -11,25 +11,26 @@ export const parseSlots = (args: Args, slotsControls: string[]) =>
       : `<template slot="${unfurledSlotName}">${args[slotName]}</template>`;
   })}`;
 
-export const renderComponent = (args: Args, context: StoryContext) => {
-  const shadowPartArgs: string[] = [];
-  const attributeControls: string[] = [];
-  const slotsControls: string[] = [];
-  for (const argKey of Object.keys(args)) {
-    switch (context.argTypes[argKey].table?.category) {
-      case 'css shadow parts':
-        shadowPartArgs.push(argKey);
-        break;
-      case 'attributes':
-        attributeControls.push(argKey);
-        break;
-      case 'slots':
-        slotsControls.push(argKey);
-        break;
+const renderComponentInternal =
+  (withCodeRoot: boolean) => (args: Args, context: StoryContext) => {
+    const shadowPartArgs: string[] = [];
+    const attributeControls: string[] = [];
+    const slotsControls: string[] = [];
+    for (const argKey of Object.keys(args)) {
+      switch (context.argTypes[argKey].table?.category) {
+        case 'css shadow parts':
+          shadowPartArgs.push(argKey);
+          break;
+        case 'attributes':
+          attributeControls.push(argKey);
+          break;
+        case 'slots':
+          slotsControls.push(argKey);
+          break;
+      }
     }
-  }
-  return html`
-  <div id="code-root">
+    return html`
+  ${withCodeRoot ? html`<div id="code-root"></div>` : ''}
     <style>
         ${unsafeStatic(
           shadowPartArgs
@@ -47,5 +48,8 @@ export const renderComponent = (args: Args, context: StoryContext) => {
       <${unsafeStatic(context.componentId)}	${unsafeStatic(attributeControls.map((arg) => `${unfurlArg(arg)}="${args[arg]}"`).join('\n'))}>
         ${unsafeStatic(parseSlots(args, slotsControls))}
       </${unsafeStatic(context.componentId)}>
-  </div>`;
-};
+  ${withCodeRoot ? html`</div>` : ''}`;
+  };
+
+export const renderComponent = renderComponentInternal(true);
+export const renderComponentWithoutCodeRoot = renderComponentInternal(false);